+++ inst/tinytest/test-ab_selectors.R
@@ -53,6 +53,9 @@ expect_equal(ncol(example_isolates[, tetracyclines(), drop = FALSE]), 3, toleran
 expect_equal(ncol(example_isolates[, trimethoprims(), drop = FALSE]), 2, tolerance = 0.5)
 expect_equal(ncol(example_isolates[, ureidopenicillins(), drop = FALSE]), 1, tolerance = 0.5)
 
+expect_message(carbapenems())
+expect_error(administrable_per_os())
+
 # Examples:
 
 # select columns 'mo', 'AMK', 'GEN', 'KAN' and 'TOB'